Add PWM motor ramping, DEFCON-scaled klaxon, and rollover fixes
Motor:
- New DoorMotor class owning both speed controller inputs (pins 9/8).
  Ramps rest-to-full over MOTOR_RAMP_MS and always passes through rest
  when reversing, so the door no longer jerks on a direction change.
- Button release ramps down; limit switches halt immediately, since
  coasting there would drive the door into the stop.

Klaxon / DEFCON:
- DEFCON is now an integer 0-4 derived from a single tunable interval
  (DEFCON_INTERVAL_MS), replacing the fixed 10 minute float timeout.
  Accumulation caps at DEFCON 4, which then holds indefinitely.
- The klaxon trigger is held low for (DEFCON + 1) klaxon loops, and
  SOUND_NOMINAL sounds instead at DEFCON 0. The sequence runs to
  completion and resets the ladder, so button release is a no-op.
- The level bar is driven by the same integer via the new
  LedLevelIndicator::setLitCount, so the display cannot drift out of
  step with the number of times the klaxon sounds.

Fixes:
- The inside shop door button is edge triggered via its pressed
  callback. Polling it fired the error sound every pass once the door
  left a limit switch while the button was still held.
- loop() reads millis() once and relies on unsigned wraparound rather
  than a special case that was both unnecessary and off by one.
- Motion widened to unsigned long. Holding millis() in an unsigned int
  truncated it to 16 bits on AVR and wrapped every ~65 seconds.
- Dropped two Button::deactivate() calls that did not compile; the
  airlock reed switches are inputs with no LED to turn off.

Verified: compiles clean for arduino:avr:mega:cpu=atmega2560.
Not yet tested on hardware.

+++ SciFiDoor/Motion.hpp
@@ -9,5 +9,5 @@ namespace Motion {
 
   boolean getMotion();
 
-  unsigned int getTimeSinceLastMotion();
+  unsigned long getTimeSinceLastMotion();
 }
